Rename text editor duck
