+++ app/webpack/observations/shared/ducks/text_editor.js
